reorganize write
